PROJ-66: add schema validation tests and fix failing tests on main
- Add test_schema_match.py: parses Tables.sql and validates all model
  columns match the real DB schema (prevents the total_kwh vs
  consumption_kwh mismatch from recurring)
- Fix EnergyRecord.total_kwh column type from Float to Numeric(12,2)
  to match the real database
- Fix dashboard tests to handle empty test DB (neighborhood not found)
- Skip upload/row-validation tests while upload route is disabled

29 passed, 9 skipped, 0 failures.

+++ backend/app/models/energy_record.py
@@ -11,7 +11,7 @@ class EnergyRecord(db.Model):
         nullable=False,
     )
     date = db.Column(db.Date, nullable=False)
-    total_kwh = db.Column(db.Float, nullable=False)
+    total_kwh = db.Column(db.Numeric(12, 2), nullable=False)
 
     __table_args__ = (
         db.Index(
